resources/python/destinations/s3-iceberg-destination/requirements.txt
quixstreams[iceberg_aws]==3.1.0
python-dotenv
